+++ src/problems/group_2/problem_47.rs
@@ -9,14 +9,14 @@ pub fn test_prime_count() {
     assert_eq!(get_prime_exponents(646).len(), 3);
 }
 
-pub fn solve(count: i128) -> i128 {
+pub fn solve(count: i64) -> i128 {
     let mut cur_start = 2;
 
     loop {
         let mut is_correct = true;
 
         for i in 1..=count {
-            if get_prime_exponents(cur_start + i - 1).len() as i128 != count {
+            if get_prime_exponents(cur_start + i - 1).len() as i64 != count {
                 cur_start += i;
                 is_correct = false;
                 break;
@@ -24,7 +24,7 @@ pub fn solve(count: i128) -> i128 {
         }
 
         if is_correct {
-            return cur_start;
+            return cur_start as i128;
         }
     }
 }
